Simplify approach to one single function
- keep that function in a file that is already synchronized with WordPress.com
- include both wpcom and Jetpack tracking approaches in that function.
- remove the dev mode check since this will be addressed in #13711
